+++ components/GoalInput.js
@@ -1,0 +1,44 @@
+import { useState } from 'react';
+import { View, TextInput, Button, StyleSheet } from 'react-native';
+
+function GoalInput(props) {
+    const [enteredGoalText, setEnteredGoalText] = useState('');
+
+    function goalInputHandler(enteredText) {
+        setEnteredGoalText(enteredText);
+    };
+
+    function addGoalHandler() {
+        props.onAddGoal(enteredGoalText);
+        setEnteredGoalText('');
+    };
+
+    return (
+        <View style={styles.inputContainer}>
+            <TextInput style={styles.textInput} placeholder="Your current goal!" onChangeText={goalInputHandler}/>
+            <Button title='Add goal' onPress={addGoalHandler}/>
+        </View>
+    )
+}
+
+export default GoalInput;
+
+const styles = StyleSheet.create({
+  inputContainer: {
+    flex: 1,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: 24,
+    borderBottomWidth: 1,
+    borderBottomColor: '#cccccc'
+  },
+  textInput: {
+    borderWidth: 1,
+    borderColor: '#cccccc',
+    width: '75%',
+    marginRight: 8,
+    paddingHorizontal: 8,
+    paddingVertical: 8
+  }
+});
